refactoring Composite objects; adding Printable flag to the graphics object; cleaning up unused attributes; adding const bit array; adding -a (align) option to 'Label' macro;

+++ gxlib/xatom.h
@@ -61,7 +61,7 @@ public:
   };
   class Settings;
 private:
-  short FDrawStyle, FRadius;
+  int FDrawStyle, FRadius;
   GLuint ActualSphere;
   // picks the correct sphere for rendering
   void InitActualSphere();
